#148 Move endpoint to new namespace and route, add client api resource file
